Fix license
I previously didn't include the year and name.

I'm not going to include my full legal name here. https://opensource.stackexchange.com/questions/4273/which-author-copyright-holder-type-is-allowed-in-the-mit-license-pseudonym-c seems to indicate it's OK to use a psuedonym under UK and US law, and if it's not, email me/open an issue and I'll work something out.